visualize

+++ demo_patch.py
@@ -223,7 +223,7 @@ for id, path in enumerate(fpaths):
     if not os.path.exists(save_folder):
             os.makedirs(save_folder)
 
-    x_adv = x_adv.astype(np.uint8)
+    # x_adv = x_adv.astype(np.uint8)
     # print(x_adv.dtype)
     img = T.ToPILImage()(x_adv[0])
     print(save_name)
